implicit-casts:false in flutter/lib/src/physics

+++ packages/flutter/lib/src/physics/friction_simulation.dart
@@ -65,7 +65,7 @@ class FrictionSimulation extends Simulation {
   // Solving for D given x(time) is trickier. Algebra courtesy of Wolfram Alpha:
   // x1 = x0 + (v0 * D^((log(v1) - log(v0)) / log(D))) / log(D) - v0 / log(D), find D
   static double _dragFor(double startPosition, double endPosition, double startVelocity, double endVelocity) {
-    return math.pow(math.e, (startVelocity - endVelocity) / (startPosition - endPosition));
+    return math.pow(math.e, (startVelocity - endVelocity) / (startPosition - endPosition)) as double;
   }
 
   @override
@@ -116,7 +116,7 @@ class BoundedFrictionSimulation extends FrictionSimulation {
 
   @override
   double x(double time) {
-    return super.x(time).clamp(_minX, _maxX);
+    return super.x(time).clamp(_minX, _maxX) as double;
   }
 
   @override
